Add files via upload
I have worked on a advertising data set from Kaggle, indicating whether a particular internet user clicked on a fake advertisement or not. Later have created a model that will predict whether or not they will click on an ad based on the features of that user.

This data set contains the following features:

Daily Time Spent on Site: Time spent by user in minutes
Age: User's age
Area Income: Avg. Income of geographical area of user
Daily Internet Usage: Avg. minutes a day the user is on the internet
Ad Topic Line: Caption of the advertisement
City: City of the user
Male: whether user is male or not
Country: Country of the user
Timestamp: Time at which user clicked on Ad or closed window
Clicked on Ad: 0 or 1 indicated clicking on Ad (This is the label based which we will be predicting in the model)